Added Alert. Removed copy-paste.
