+++ components/email_aliases/resources/content/email_aliases_manage_page.tsx
@@ -7,6 +7,7 @@ import { font, typography } from '@brave/leo/tokens/css/variables'
 import { Introduction } from './email_aliases_introduction'
 import { AliasList } from './email_aliases_list'
 import * as React from 'react'
+import Alert from '@brave/leo/react/alert'
 import Col from './styles/Col'
 import styled from 'styled-components'
 import {
@@ -35,22 +36,27 @@ const BraveAccountSignIn = () => {
 
 export const ManagePage = ({
   aliasesState,
+  aliasesError,
   authState,
   emailAliasesService,
 }: {
   aliasesState: Alias[]
+  aliasesError: string | null
   authState: AuthState
   emailAliasesService: EmailAliasesServiceInterface
 }) => (
   <PageCol>
     <Introduction />
     <BraveAccountSignIn />
-    {authState.status === AuthenticationStatus.kAuthenticated && (
-      <AliasList
-        aliases={aliasesState}
-        authEmail={authState.email}
-        emailAliasesService={emailAliasesService}
-      />
-    )}
+    {authState.status === AuthenticationStatus.kAuthenticated
+      && (aliasesError ? (
+        <Alert type='error'>{aliasesError}</Alert>
+      ) : (
+        <AliasList
+          aliases={aliasesState}
+          authEmail={authState.email}
+          emailAliasesService={emailAliasesService}
+        />
+      ))}
   </PageCol>
 )
