+++ components/email_aliases/resources/content/use_email_aliases_observer.ts
@@ -1,0 +1,59 @@
+// Copyright (c) 2026 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
+import * as React from 'react'
+import {
+  AliasesUpdate,
+  AuthenticationStatus,
+  AuthState,
+  Alias,
+  EmailAliasesServiceObserverInterface,
+} from 'gen/brave/components/email_aliases/email_aliases.mojom.m'
+
+/**
+ * Subscribes to EmailAliasesService observer callbacks and exposes auth state,
+ * alias list, and list-load error. Pass a stable `bindObserver` from the host
+ * (e.g. one created when the Mojo pipes are set up).
+ */
+export function useEmailAliasesObserver(
+  bindObserver: (observer: EmailAliasesServiceObserverInterface) => () => void,
+) {
+  const [authState, setAuthState] = React.useState<AuthState>({
+    status: AuthenticationStatus.kStartup,
+    email: '',
+  })
+  const [aliasesState, setAliasesState] = React.useState<Alias[]>([])
+  const [aliasesError, setAliasesError] = React.useState<string | null>(null)
+
+  React.useEffect(() => {
+    // Track auth locally so we do not apply alias updates while logged out.
+    let status: AuthenticationStatus = AuthenticationStatus.kStartup
+    const observer: EmailAliasesServiceObserverInterface = {
+      onAliasesUpdated: (update: AliasesUpdate) => {
+        if (status !== AuthenticationStatus.kAuthenticated) {
+          return
+        }
+        if (update.aliases) {
+          setAliasesState(update.aliases!)
+          setAliasesError(null)
+        } else {
+          setAliasesState([])
+          setAliasesError(update.error!)
+        }
+      },
+      onAuthStateChanged: (state: AuthState) => {
+        status = state.status
+        setAuthState(state)
+        if (status !== AuthenticationStatus.kAuthenticated) {
+          setAliasesState([])
+          setAliasesError(null)
+        }
+      },
+    }
+    return bindObserver(observer)
+  }, [])
+
+  return { authState, aliasesState, aliasesError }
+}
